added driver code and verified RTC

+++ drivers/NEOM8N/neom8n.h
@@ -4,4 +4,4 @@
 #include "hardware/dma.h"
 
 
-void neom8n_init(uart_inst_t *uart, uint baudrate, uint8_t *rx_buf, size_t buf_size, int *dma_chan) ;
+void neom8n_init(uart_inst_t *uart, uint8_t baudrate, uint8_t *rx_buf, size_t buf_size, int *dma_chan) ;
